bindings: counter: extensibility with channel specifier
It was not yet possible to specify in the device tree which channel a
counter instance should use. This made drivers using counters less
portable. This is now fixed by adding a new counter.yaml file that
adds the #counter-cells property. This file can be included in the
counter binding.

Signed-off-by: Jaro Van Landschoot <jaro.vanlandschoot@basalte.be>